+++ tests/test_core/test_readers/test_read_labspec.py
@@ -4,7 +4,7 @@
 import pytest
 
 import spectrochempy as scp
-from spectrochempy.utils import show
+from pathlib import Path
 
 RAMANDIR = scp.preferences.datadir / "ramandata"
 
@@ -15,8 +15,24 @@ RAMANDIR = scp.preferences.datadir / "ramandata"
 )
 def test_read_labspec():
 
+    # single file
     nd = scp.read_labspec("Activation.txt", directory=RAMANDIR)
     assert nd.shape == (532, 1024)
 
+    # with read_dir
     nd = scp.read_dir(directory=RAMANDIR / "subdir")
     assert nd.shape == (6, 1024)
+
+    # empty txt file
+    Path("i_am_empty.txt").touch()
+    f = Path("i_am_empty.txt")
+    nd = scp.read_labspec(f)
+    f.unlink()
+    assert nd is None
+
+    # non labspec txt file
+    f = Path("i_am_not_labspec.txt")
+    f.write_text("blah")
+    nd = scp.read_labspec(f)
+    f.unlink()
+    assert nd is None