+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 
 * Issue #360 : write a dataset with a specified filename do not open a dialog except if the file already exists and if confirm=True is set.
 
+* read and read_dir: now ignore non readable files
+
+* read_labspec: now ignore non-labspec .txt files
+
 ## Version 0.3.1
 
 **NEW FEATURES**